Bug 1766377 - Fix sign-compare warnings related to IDL constants. r=anti-tracking-reviewers,necko-reviewers,kershaw,timhuang
dom/network/TCPSocket.cpp(973,38): error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'nsISocketTransport::(unnamed enum at obj-aarch64-pc-mingw32/dist/include/nsISocketTransport.h:102:3)' [-Werror,-Wsign-compare]
  if (static_cast<uint32_t>(aStatus) !=
      ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~ ^
dom/security/ReferrerInfo.cpp(216,26): error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'nsIWebProgressListener::(unnamed enum at obj-aarch64-pc-mingw32/dist/include/nsIWebProgressListener.h:41:3)' [-Werror,-Wsign-compare]
          rejectedReason !=
          ~~~~~~~~~~~~~~ ^
toolkit/components/antitracking/StorageAccess.cpp(175,23): error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'nsIWebProgressListener::(unnamed enum at obj-aarch64-pc-mingw32/dist/include/nsIWebProgressListener.h:41:3)' [-Werror,-Wsign-compare]
  if (aRejectedReason ==
      ~~~~~~~~~~~~~~~ ^
toolkit/components/antitracking/StorageAccess.cpp(415,26): error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'nsIWebProgressListener::(unnamed enum at obj-aarch64-pc-mingw32/dist/include/nsIWebProgressListener.h:41:3)' [-Werror,-Wsign-compare]
         aRejectedReason ==
         ~~~~~~~~~~~~~~~ ^
toolkit/components/antitracking/StorageAccess.cpp(428,26): error: comparison of integers of different signs: 'uint32_t' (aka 'unsigned int') and 'nsIWebProgressListener::(unnamed enum at obj-aarch64-pc-mingw32/dist/include/nsIWebProgressListener.h:41:3)' [-Werror,-Wsign-compare]
  return aRejectedReason ==
         ~~~~~~~~~~~~~~~ ^

Differential Revision: https://phabricator.services.mozilla.com/D144693

+++ toolkit/components/antitracking/StorageAccess.cpp
@@ -165,15 +165,18 @@ static StorageAccess InternalStorageAllowedCheck(
 
   // We want to have a partitioned storage only for trackers.
   if (aRejectedReason ==
-          nsIWebProgressListener::STATE_COOKIES_BLOCKED_TRACKER ||
+          static_cast<uint32_t>(
+              nsIWebProgressListener::STATE_COOKIES_BLOCKED_TRACKER) ||
       aRejectedReason ==
-          nsIWebProgressListener::STATE_COOKIES_BLOCKED_SOCIALTRACKER) {
+          static_cast<uint32_t>(
+              nsIWebProgressListener::STATE_COOKIES_BLOCKED_SOCIALTRACKER)) {
     return StorageAccess::ePartitionTrackersOrDeny;
   }
 
   // We want to have a partitioned storage for all third parties.
   if (aRejectedReason ==
-      nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN) {
+      static_cast<uint32_t>(
+          nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN)) {
     return StorageAccess::ePartitionForeignOrDeny;
   }
 
@@ -409,11 +412,14 @@ bool ShouldPartitionStorage(StorageAccess aAccess) {
 
 bool ShouldPartitionStorage(uint32_t aRejectedReason) {
   return aRejectedReason ==
-             nsIWebProgressListener::STATE_COOKIES_BLOCKED_TRACKER ||
+             static_cast<uint32_t>(
+                 nsIWebProgressListener::STATE_COOKIES_BLOCKED_TRACKER) ||
          aRejectedReason ==
-             nsIWebProgressListener::STATE_COOKIES_BLOCKED_SOCIALTRACKER ||
+             static_cast<uint32_t>(
+                 nsIWebProgressListener::STATE_COOKIES_BLOCKED_SOCIALTRACKER) ||
          aRejectedReason ==
-             nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN;
+             static_cast<uint32_t>(
+                 nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN);
 }
 
 bool StoragePartitioningEnabled(StorageAccess aAccess,
@@ -426,7 +432,8 @@ bool StoragePartitioningEnabled(StorageAccess aAccess,
 bool StoragePartitioningEnabled(uint32_t aRejectedReason,
                                 nsICookieJarSettings* aCookieJarSettings) {
   return aRejectedReason ==
-             nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN &&
+             static_cast<uint32_t>(
+                 nsIWebProgressListener::STATE_COOKIES_PARTITIONED_FOREIGN) &&
          aCookieJarSettings->GetCookieBehavior() ==
              nsICookieService::BEHAVIOR_REJECT_TRACKER_AND_PARTITION_FOREIGN;
 }
